+++ CHANGELOG.md
@@ -21,6 +21,8 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 - Fonts are now built-in and automatically managed, remove any `~@eightshift/ui-components/dist/assets/fonts.css` imports
 - `ColorPicker` now supports adding custom options below colors via the `extraOptions` prop
 - Added 2 new UI icons
+- `InputField`s can now be `inline`
+- Simplified `MatrixAlign` (`popoverPosition` prop is no longer supported)
 
 ## [1.9.1] - 2024-12-17
 - Downgraded from React 19 to React 18.